requirements.txt
asgiref==3.9.2
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
uritemplate==4.2.0
gunicorn==23.0.0
dj-database-url==3.0.1
